Fix circularity between llnl.util.filesystem and llnl.util.symlink
