+++ lib/spack/spack/test/cmd/view.py
@@ -11,7 +11,7 @@ import pytest
 import spack.concretize
 import spack.util.spack_yaml as s_yaml
 from spack.installer import PackageInstaller
-from spack.llnl.util.symlink import _windows_can_symlink
+from spack.llnl.util.filesystem import _windows_can_symlink
 from spack.main import SpackCommand
 
 extensions = SpackCommand("extensions")
